feat: 씬 연결
