+++ SuspiciousDiary/Assets/Dialogue/Ep1_ECC/Ep1_4.asset
@@ -13,5 +13,6 @@ MonoBehaviour:
   m_Name: Ep1_4
   m_EditorClassIdentifier: 
   lines:
-  - "\uC228\uC5B4\uC788\uB358 \uC601\uD654\uC758 \uBD80\uC81C\uAC00 \uB098\uD0C0\uB0AC\uC5B4"
+  - "\uC228\uC5B4\uC788\uB358 \uC601\uD654\uC758 \uBD80\uC81C\uAC00 \uB098\uD0C0\uB0AC\uC5B4!   
+    \uC0AC\uB77C\uC9C4 \uC911\uC559\uB3C4\uC11C\uAD00?"
   - "\uC774\uAC83\uB3C4 \uBB34\uC5B8\uAC00 \uAC00\uB9AC\uD0A4\uACE0 \uC788\uB294\uAC74\uAC00?"
