+++ SuspiciousDiary/Assets/Dialogue/Ep2_library/Ep2_2.asset
@@ -15,4 +15,3 @@ MonoBehaviour:
   lines:
   - "\uCDA9\uBD84\uD788 \uCDA9\uC804\uB410\uB124.. \uB2E4\uD589\uC774\uB2E4"
   - "\uC557! \uCC44\uD50C\uC744 \uC78A\uACE0 \uC788\uC5C8\uC796\uC544..!"
-  - "<color=red>\uC54C\uB78C\uC744 \uD074\uB9AD\uD558\uC2DC\uC624</color>"
