+++ SuspiciousDiary/Assets/Dialogue/Ep2_library/Ep2_3.asset
@@ -1,0 +1,16 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: bb4d5072a9e8e4ee5bd9e055d3232a46, type: 3}
+  m_Name: Ep2_3
+  m_EditorClassIdentifier: 
+  lines:
+  - "<color=red>\uC54C\uB78C\uC744 \uD074\uB9AD\uD558\uC2DC\uC624</color>"
