+++ SuspiciousDiary/Assets/Dialogue/Ep2_library/Ep2_3.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5d74c6aa643bc9e46bf0b4d284e1a67e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
