+++ SuspiciousDiary/Assets/Editor/Vuforia/ImageTargetTextures/SuspiciousDiary/Asan123_scaled.jpg.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b824ee616fc94608ad7c07e072f4d8bb
+guid: 44af5ff6b2be40fcaaac2af4ba015838
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
