+++ SuspiciousDiary/Assets/Editor/Vuforia/ImageTargetTextures/SuspiciousDiary/RestaurantPoster_scaled.jpg.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 58a63a1ff03d415892dd9c401ff9d791
+guid: 851ad58f526e45eb9f77e55a63868b33
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
