+++ SuspiciousDiary/Assets/Editor/Vuforia/ImageTargetTextures/SuspiciousDiary/WideFlower_scaled.jpg.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7242397323874010a97b0bfc04ed022d
+guid: f54d849e85304b3e974cd217e563fe08
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
