+++ SuspiciousDiary/Assets/Editor/Vuforia/ImageTargetTextures/SuspiciousDiary/arthouse_momo_scaled.jpg.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 23ed7700d8fd4f6f8dc9a08cb3f19552
+guid: db909baaa4574d19a922f194bff1beed
 TextureImporter:
   internalIDToNameTable: []
   externalObjects: {}
