--- SuspiciousDiary/Assets/Scenes/Auditorium.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9dc453667b341a84bbff0d3ad9cded50
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
