+++ SuspiciousDiary/Assets/Scenes/Episode2_library.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3a16705a62c024ad9b6271c86647d095
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
