+++ SuspiciousDiary/Assets/Scenes/Episode3_Auditorium.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9dc453667b341a84bbff0d3ad9cded50
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
