+++ SuspiciousDiary/Assets/Scenes/Episode4_AsanEngineering.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b3ffbffa0c0bdba449d44cabb1a5f14a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
