+++ SuspiciousDiary/Assets/Scripts/ARSceneChanger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aed7906dda818fd4c84da036971dc3ca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
