+++ SuspiciousDiary/Assets/Scripts/EndingManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 41bfb156cedbd8a47bbfbd6666333ed4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
