+++ SuspiciousDiary/Assets/Scripts/MinigameResult.cs
@@ -1,0 +1,5 @@
+
+public static class MinigameResult
+{
+    public static bool isSuccess = true;
+}
