+++ SuspiciousDiary/Assets/Scripts/MinigameResult.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a742abfb425a8bb4aadb8a617eba02d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
