+++ SuspiciousDiary/Assets/Scripts/SceneChanger.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class SceneChanger : MonoBehaviour
+{
+    public string sceneToLoad;
+    public void ChangeScene()
+    {
+        SceneManager.LoadScene(sceneToLoad);
+    }
+}
