+++ SuspiciousDiary/Assets/Scripts/SceneChanger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c93924fd907587e4086e7008d0cf9fcc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
