--- SuspiciousDiary/Assets/Scripts/StageManager_Ep1.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cd3061d1d77ad8447ab6e2f61e4be774
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
