+++ SuspiciousDiary/Assets/StreamingAssets/Vuforia/SuspiciousDiary.dat.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: af1cbaf667e443b49bcfc07a5845787f
+guid: 1b4bc6089e2140ccac9cbb7b4e1af76d
