+++ SuspiciousDiary/Assets/StreamingAssets/Vuforia/SuspiciousDiary.xml.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: c69e58fcd4bc4987861ae288bf2f01ea
+guid: edd686a862a34d65a59b92c1230818d1
